+++ package.json
@@ -21,6 +21,8 @@
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "express": "^4.17.1",
+    "mongodb": "^3.5.6",
+    "mongoose": "^5.9.9",
     "node": "^13.13.0",
     "nodemon": "^2.0.3",
     "path": "^0.12.7",
